updated install, all_js generation, paster deamon stopping commands

+++ bqcore/bq/util/copylink.py
@@ -2,7 +2,12 @@ import os,sys
 import shutil
 import logging
 
+from .mkdir import _mkdir
+from .io_misc import dolink
+
 log = logging.getLogger('bq.util.copylink')
+
+
 
 def copy_link (*largs):
     "Copy or make a hard link"
@@ -21,25 +26,51 @@ def copy_link (*largs):
             if os.path.exists(dest):
                 log.info ("Found existing file %s: removing .." % dest)
                 os.unlink (dest)
-            os.link(f, dest)
+
+            #os.link(f, dest)
+            dolink(f, dest)
+
         except (OSError, AttributeError), e:
-            log.info ("Problem in link %s .. trying copy" % e)
+            log.info ("Problem in link %s... trying copy" % e)
             if os.path.isdir(f):
                 shutil.copytree(f, dest)
             else:
                 shutil.copy2(f, dest)
 
+if os.name != 'nt':
+    def copy_symlink (source, dest):
+        "Copy or make a symlink"
+        try:
+            os.symlink(source, dest)
+        except (OSError, AttributeError), e:
+            log.info ("Problem in link %s... trying copy" % e)
+            if os.path.isdir(source):
+                shutil.copytree(source, dest)
+            else:
+                shutil.copy2(source, dest)
+else:
+    def copy_symlink (source, dest):
+        #source = os.path.abspath(source)
+        #dest = os.path.abspath(dest)
 
+        "Copy or make a symlink"
+        if not os.path.exists(source):
+            log.info('copy_symlink: source path does not exist "%s", skipping...'%(source))
+            return
 
+        #log.info('copy_symlink: "%s" -> "%s"'%(source, dest))
 
-def copy_symlink (source, dest):
-    "Copy or make a symlink"
-    try:
-        os.symlink(source, dest)
-    except (OSError, AttributeError), e:
-        log.info ("Problem in link %s .. trying copy" % e)
         if os.path.isdir(source):
-            shutil.copytree(source, dest)
+            _mkdir(dest)
+            # walk directory and call copy_symlink recursively on its children
+            fs = os.listdir(source)
+            for f in fs:
+                copy_symlink ( os.path.join(source, f), os.path.join(dest, f))
         else:
-            shutil.copy2(source, dest)
+            # try linking
+            try:
+                dolink(source, dest)
+            except (OSError, AttributeError), e:
+                log.info ("Problem in link %s... trying copy" % e)
+                shutil.copy2(source, dest)
